pyproject.toml
[project]
name = "mcp-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.43.1",
    "black>=24.10.0",
    "langchain-openai>=0.3.1",
    "langgraph>=0.2.65",
    "langsmith>=0.2.11",
    "mcp>=1.2.0",
    "openai>=1.59.9",
    "python-dotenv>=1.0.1",
]
